+++ core/Azure.Mcp.Core/src/Services/Telemetry/TelemetryService.cs
@@ -1,10 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
 using Azure.Mcp.Core.Areas.Server.Options;
 using Azure.Mcp.Core.Configuration;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using ModelContextProtocol.Protocol;
 using static Azure.Mcp.Core.Services.Telemetry.TelemetryConstants;
@@ -16,16 +18,28 @@ namespace Azure.Mcp.Core.Services.Telemetry;
 /// </summary>
 internal class TelemetryService : ITelemetryService
 {
-    private readonly bool _isEnabled;
-    private readonly List<KeyValuePair<string, object?>> _tagsList;
     private readonly IMachineInformationProvider _informationProvider;
-    private readonly TaskCompletionSource _isInitialized = new TaskCompletionSource();
+    private readonly bool _isEnabled;
+    private readonly ILogger<TelemetryService> _logger;
+    private readonly List<KeyValuePair<string, object?>> _tagsList;
+    private readonly SemaphoreSlim _initalizeLock = new(1);
+
+    /// <summary>
+    /// Task created on the first invocation of <see cref="InitializeAsync"/>.
+    /// This is saved so that repeated invocations will see the same exception
+    /// as the first invocation.
+    /// </summary>
+    private Task? _initalizationTask = null;
+
+    private bool _initializationSuccessful;
+    private bool _isInitialized;
 
     internal ActivitySource Parent { get; }
 
     public TelemetryService(IMachineInformationProvider informationProvider,
         IOptions<AzureMcpServerConfiguration> options,
-        IOptions<ServiceStartOptions>? serverOptions)
+        IOptions<ServiceStartOptions>? serverOptions,
+        ILogger<TelemetryService> logger)
     {
         _isEnabled = options.Value.IsTelemetryEnabled;
         _tagsList = new List<KeyValuePair<string, object?>>()
@@ -40,29 +54,39 @@ internal class TelemetryService : ITelemetryService
 
         Parent = new ActivitySource(options.Value.Name, options.Value.Version, _tagsList);
         _informationProvider = informationProvider;
-
-        Task.Factory.StartNew(InitializeTagList);
+        _logger = logger;
     }
 
     /// <summary>
     /// TESTING PURPOSES ONLY: Gets the default tags used for telemetry.
     /// </summary>
-    internal async Task<IReadOnlyList<KeyValuePair<string, object?>>> GetDefaultTags()
+    internal IReadOnlyList<KeyValuePair<string, object?>> GetDefaultTags()
     {
-        await _isInitialized.Task;
+        if (!_isEnabled)
+        {
+            return [];
+        }
+
+        CheckInitialization();
         return _tagsList.ToImmutableList();
     }
 
-    public ValueTask<Activity?> StartActivity(string activityId) => StartActivity(activityId, null);
+    /// <summary>
+    /// <inheritdoc/>
+    /// </summary>
+    public Activity? StartActivity(string activityId) => StartActivity(activityId, null);
 
-    public async ValueTask<Activity?> StartActivity(string activityId, Implementation? clientInfo)
+    /// <summary>
+    /// <inheritdoc/>
+    /// </summary>
+    public Activity? StartActivity(string activityId, Implementation? clientInfo)
     {
         if (!_isEnabled)
         {
             return null;
         }
 
-        await _isInitialized.Task;
+        CheckInitialization();
 
         var activity = Parent.StartActivity(activityId);
 
@@ -88,21 +112,79 @@ internal class TelemetryService : ITelemetryService
     {
     }
 
-    private async Task InitializeTagList()
+    /// <summary>
+    /// <inheritdoc/>
+    /// </summary>
+    public async Task InitializeAsync()
     {
-        try
+        if (!_isEnabled)
         {
-            var macAddressHash = await _informationProvider.GetMacAddressHash();
-            var deviceId = await _informationProvider.GetOrCreateDeviceId();
-
-            _tagsList.Add(new(TagName.MacAddressHash, macAddressHash));
-            _tagsList.Add(new(TagName.DevDeviceId, deviceId));
-
-            _isInitialized.SetResult();
+            return;
         }
-        catch (Exception ex)
+
+        // Quick check if initialization already happened. Avoids
+        // trying to get the lock.
+        if (_initalizationTask == null)
         {
-            _isInitialized.SetException(ex);
+            // Get async lock for starting initialization
+            await _initalizeLock.WaitAsync();
+
+            try
+            {
+                // Check after acquiring lock to ensure we honor work
+                // started while we were waiting.
+                if (_initalizationTask == null)
+                {
+                    _initalizationTask = InnerInitializeAsync();
+                }
+            }
+            finally
+            {
+                _initalizeLock.Release();
+            }
         }
+
+        // Await the response of the initialization work regardless of if
+        // we or another invocation created the Task representing it. All
+        // awaiting on this will give the same result to ensure idempotency.
+        await _initalizationTask;
+
+        async Task InnerInitializeAsync()
+        {
+            try
+            {
+                var macAddressHash = await _informationProvider.GetMacAddressHash();
+                var deviceId = await _informationProvider.GetOrCreateDeviceId();
+
+                _tagsList.Add(new(TagName.MacAddressHash, macAddressHash));
+                _tagsList.Add(new(TagName.DevDeviceId, deviceId));
+
+                _initializationSuccessful = true;
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Error occurred initializing telemetry service.");
+                throw;
+            }
+            finally
+            {
+                _isInitialized = true;
+            }
+        }
+    }
+
+    private void CheckInitialization()
+    {
+        if (!_isInitialized)
+        {
+            throw new InvalidOperationException(
+                $"Telemetry service has not been initialized. Use {nameof(InitializeAsync)}() before any other operations.");
+        }
+
+        if (!_initializationSuccessful)
+        {
+            throw new InvalidOperationException("Telemetry service was not successfully initialized. Check logs for initialization errors.");
+        }
+
     }
 }